+++ src/test/converter2/renderer/gh2995.ts
@@ -1,0 +1,3 @@
+export interface gh2995 {
+    optionalMethod?(filter: string, args: string[]): any;
+}