+++ src/lib/output/themes/default/partials/member.declaration.tsx
@@ -1,8 +1,27 @@
-import type { DeclarationReflection } from "../../../../models/index.js";
+import type { DeclarationReflection } from "#models";
 import { JSX } from "#utils";
 import { FormattedCodeBuilder, FormattedCodeGenerator, type FormatterNode, Wrap } from "../../../formatter.js";
 import { hasTypeParameters } from "../../lib.js";
 import type { DefaultThemeRenderContext } from "../DefaultThemeRenderContext.js";
+
+function shouldRenderDefaultValue(props: DeclarationReflection) {
+    const defaultValue = props.defaultValue;
+
+    if (defaultValue === undefined) {
+        return false;
+    }
+
+    /** Fix for #2717. If type is the same as value the default value is omitted */
+    if (props.type && props.type.type === "literal") {
+        const reflectionTypeString = props.type.toString();
+
+        if (reflectionTypeString === defaultValue.toString()) {
+            return false;
+        }
+    }
+
+    return true;
+}
 
 export function memberDeclaration(context: DefaultThemeRenderContext, props: DeclarationReflection) {
     const builder = new FormattedCodeBuilder(context.router, context.model);
@@ -11,31 +30,15 @@ export function memberDeclaration(context: DefaultThemeRenderContext, props: Dec
     const generator = new FormattedCodeGenerator(context.options.getValue("typePrintWidth"));
     generator.node({ type: "nodes", content }, Wrap.Detect);
 
-    /** Fix for #2717. If type is the same as value the default value is omitted */
-    function shouldRenderDefaultValue() {
-        if (props.type && props.type.type === "literal") {
-            const reflectionTypeString = props.type.toString();
-
-            const defaultValue = props.defaultValue;
-
-            if (defaultValue === undefined || reflectionTypeString === defaultValue.toString()) {
-                return false;
-            }
-        }
-        return true;
-    }
-
     return (
         <>
             <div class="tsd-signature">
                 {generator.toElement()}
-                {!!props.defaultValue && shouldRenderDefaultValue() && (
-                    <>
-                        <span class="tsd-signature-symbol">
-                            {" = "}
-                            {props.defaultValue}
-                        </span>
-                    </>
+                {shouldRenderDefaultValue(props) && (
+                    <span class="tsd-signature-symbol">
+                        {" = "}
+                        {props.defaultValue}
+                    </span>
                 )}
             </div>